seperates modeling and simulation code into seperate modules
